backend/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
pytesseract==0.3.10
pillow==10.3.0
openpyxl==3.1.2
python-multipart==0.0.9